Добавляет функционал подгрузки комментариев (мок) + рефакторинг
