+++ js/features/render-image-tiles/index.js
@@ -1,21 +1,27 @@
 import { fillDocumentFragment } from '/js/shared/fill-document-fragment';
-import { createTileImage } from '/js/features/image-tile';
-import { openImageModal, prepareImageModal} from '/js/features/image-modal';
+import { openImageModal } from '/js/shared/modal-handlers';
+import { createTileImage } from '/js/entities/photo';
+import { setModalImage } from '/js/features/set-modal-image';
+import { prepareCommentRender } from '/js/features/render-comments-section';
 
 const picturesPlace = document.querySelector('.pictures');
+
+const getClickTileHandler = (photos) => (evt) => {
+  evt.preventDefault();
+
+  const actualData = photos.find((photo) => evt.currentTarget.dataset.id === photo.id.toString());
+
+  setModalImage(actualData);
+  prepareCommentRender(actualData.comments);
+
+  openImageModal();
+};
 
 const imageTilesHandlers = (photos) => {
   const pictureEls = picturesPlace.querySelectorAll('.picture');
 
   pictureEls.forEach((pictureEl) => {
-    pictureEl.addEventListener('click', (evt) => {
-      evt.preventDefault();
-
-      const actualData = photos.find((photo) => evt.currentTarget.dataset.id === photo.id.toString());
-
-      prepareImageModal(actualData);
-      openImageModal();
-    });
+    pictureEl.addEventListener('click', getClickTileHandler(photos));
   });
 };
